Add support for external highlighter.
An example is included in the code comments for integration with
'highlight.js' (an excellent highlighter):

Flatdoc.run({
  highlight: function (code, value) {
     return hljs.highlight(lang, code).value;
  },
  ...
 });

Note that in addition to this, only the 'generic' highlighter has
been preserved.

The idea being that most people will want to link in 'highlight.js'
because it supports so many more languages.